chore: starts head -> metadata migration
