+++ package.json
@@ -24,7 +24,6 @@
     "@faceless-ui/slider": "^1.1.17",
     "@faceless-ui/window-info": "^2.1.1",
     "@next/bundle-analyzer": "^13.0.2",
-    "@next/font": "^13.0.3",
     "@stripe/react-stripe-js": "^1.16.5",
     "@stripe/stripe-js": "^1.47.0",
     "@vercel/og": "^0.0.21",
