+++ src/app/(pages)/[...slug]/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
+import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
 import { Hero } from '@components/Hero'
-import Meta from '@components/Meta'
 import { RenderBlocks } from '@components/RenderBlocks'
 import { fetchPage, fetchPages } from '../../../graphql'
 
@@ -13,7 +13,6 @@ const Page = async ({ params: { slug } }) => {
 
   return (
     <React.Fragment>
-      <Meta title={page.meta?.title} description={page.meta?.description} slug={slug} />
       <Hero page={page} />
       <RenderBlocks blocks={page.layout} />
     </React.Fragment>
@@ -28,4 +27,13 @@ export async function generateStaticParams() {
   return pages.map(({ breadcrumbs }) => ({
     slug: breadcrumbs?.[breadcrumbs.length - 1]?.url?.replace(/^\/|\/$/g, '').split('/'),
   }))
+}
+
+export async function generateMetadata({ params: { slug } }): Promise<Metadata> {
+  const page = await fetchPage(slug)
+
+  return {
+    title: page?.meta?.title,
+    description: page?.meta?.description,
+  }
 }
