+++ src/app/(pages)/blog/[slug]/client_page.tsx
