+++ src/app/(pages)/blog/[slug]/page.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
+import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
 import { fetchBlogPost, fetchPosts } from '../../../../graphql'
-import { RenderBlogPost } from './render'
+import { RenderBlogPost } from './client_page'
 
 const Post = async ({ params }) => {
   const { slug } = params
@@ -21,4 +22,22 @@ export async function generateStaticParams() {
   return posts.map(({ slug }) => ({
     slug,
   }))
+}
+
+export async function generateMetadata({ params: { slug } }): Promise<Metadata> {
+  const page = await fetchBlogPost(slug)
+
+  return {
+    title: page?.meta?.title,
+    description: page?.meta?.description,
+    openGraph: {
+      url: `/blog/${slug}`,
+      description: page?.meta?.description,
+      images: [
+        {
+          url: typeof page.image === 'object' && page.image?.url ? page.image.url : '',
+        },
+      ],
+    },
+  }
 }
