--- src/app/(pages)/blog/head.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import Meta from '@components/Meta'
-
-export default async () => {
-  return (
-    <React.Fragment>
-      <Meta
-        title="Payload CMS | Blog"
-        description="The official Payload CMS blog. Read about the product and keep up to date with new features."
-        slug="blog"
-      />
-    </React.Fragment>
-  )
-}
