+++ src/app/(pages)/blog/page.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { Metadata } from 'next'
 
 import { fetchBlogPosts } from '../../../graphql'
 import { RenderBlogArchive } from './renderBlogArchive'
@@ -9,3 +10,9 @@ const Page = async () => {
 }
 
 export default Page
+
+export const metadata: Metadata = {
+  openGraph: {
+    url: '/blog',
+  },
+}
