+++ src/app/(pages)/login/page.tsx
@@ -12,7 +12,7 @@ import { redirect, useSearchParams } from 'next/navigation'
 import { Gutter } from '@components/Gutter'
 import { Heading } from '@components/Heading'
 import { Message } from '@components/Message'
-import { cloudSlug } from '@root/app/cloud/layout'
+import { cloudSlug } from '@root/app/cloud/client_layout'
 import { useAuth } from '@root/providers/Auth'
 import canUseDom from '@root/utilities/can-use-dom'
