+++ src/app/fonts.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line camelcase
-import { Roboto_Mono } from '@next/font/google'
-import localFont from '@next/font/local'
+import { Roboto_Mono } from 'next/font/google'
+import localFont from 'next/font/local'
 
 export const robotoMono = Roboto_Mono({
   weight: ['400'],
