+++ src/app/new/(checkout)/Checkout.tsx
@@ -23,7 +23,7 @@ import { useInstallationSelector } from '@components/InstallationSelector'
 import { LoadingShimmer } from '@components/LoadingShimmer'
 import { usePlanSelector } from '@components/PlanSelector'
 import { TeamSelector } from '@components/TeamSelector'
-import { cloudSlug } from '@root/app/cloud/layout'
+import { cloudSlug } from '@root/app/cloud/client_layout'
 import { Plan, Project, Team } from '@root/payload-cloud-types'
 import { useAuth } from '@root/providers/Auth'
 import { useGlobals } from '@root/providers/Globals'
