+++ src/app/page.tsx
@@ -1,3 +1,5 @@
-import PageTemplate from './(pages)/[...slug]/page'
+import PageTemplate, { generateMetadata } from './(pages)/[...slug]/page'
 
 export default PageTemplate
+
+export { generateMetadata }
