+++ src/app/test/page.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default () => {
+  return <h1>test page</h1>
+}
