+++ src/components/cards/ProjectCard/index.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 import Link from 'next/link'
 
-import { cloudSlug } from '@root/app/cloud/layout'
+import { cloudSlug } from '@root/app/cloud/client_layout'
 import { BranchIcon } from '@root/graphics/BranchIcon'
 import { GitHubIcon } from '@root/graphics/GitHub'
 import { ArrowIcon } from '@root/icons/ArrowIcon'
